Tests/LibWeb/Text/expected/css/has-fast-reject-filter-quirks-class-selector.txt

Quirks mode: BackCompat
:has() matches case-insensitive class selector after cached miss: true
